changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11772 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Data Processing/DataExplorer/ORHeaderItem.h
@@ -48,7 +48,7 @@
 - (NSMutableArray *) items;
 - (void) setItems: (NSMutableArray *) anItems;
 - (void) addObject:(id)anObject;
-- (unsigned) count;
+- (NSUInteger) count;
 - (BOOL) isLeafNode;
-- (id) childAtIndex:(int)index;
+- (id) childAtIndex:(NSUInteger)index;
 @end
